Crée un store spécifique pour prévention.
Avec une partie commune à sécurité.
