+++ src/situations/prevention/modeles/store.js
@@ -1,0 +1,25 @@
+import {
+  CHARGEMENT,
+  ENTRAINEMENT_DEMARRE,
+  ENTRAINEMENT_FINI,
+  DEMARRE,
+  FINI
+} from 'commun/modeles/situation';
+import { creeStore as creeStoreCommun } from 'commun/modeles/store';
+
+export function creeStore () {
+  return creeStoreCommun({
+    state: {
+      zones: [],
+      fondSituation: ''
+    },
+    mutations: {
+      configureActe (state, { zones, fondSituation }) {
+        state.zones = zones;
+        state.fondSituation = fondSituation;
+      }
+    }
+  });
+}
+
+export { CHARGEMENT, ENTRAINEMENT_DEMARRE, ENTRAINEMENT_FINI, DEMARRE, FINI };
